requirements.txt
asttokens==2.0.8
backcall==0.2.0
bcrypt==4.0.1
beautifulsoup4==4.11.1
bs4==0.0.1
capstone==5.0.0rc2
certifi==2022.9.24
cffi==1.15.1
charset-normalizer==2.1.1
colored-traceback==0.3.0
contourpy==1.0.5
cryptography==38.0.1
cycler==0.11.0
debugpy==1.6.3
decorator==5.1.1
entrypoints==0.4
executing==1.1.1
fonttools==4.37.4
fuzzywuzzy==0.18.0
gdata==2.0.18
greenlet==1.1.3.post0
idna==3.4
intervaltree==3.1.0
ipykernel==6.16.0
ipython==8.5.0
jedi==0.18.1
jupyter-core==4.11.1
jupyter_client==7.4.2
kiwisolver==1.4.4
Levenshtein==0.20.7
lxml==4.9.1
Mako==1.2.3
MarkupSafe==2.1.1
matplotlib==3.6.1
matplotlib-inline==0.1.6
nest-asyncio==1.5.6
numpy==1.23.4
packaging==21.3
pandas==1.5.0
paramiko==2.11.0
parso==0.8.3
pathlib2==2.3.7.post1
pexpect==4.8.0
pickleshare==0.7.5
Pillow==9.2.0
plumbum==1.8.0
prompt-toolkit==3.0.31
psutil==5.9.2
psycopg2-binary==2.9.4
ptyprocess==0.7.0
pure-eval==0.2.2
pwn==1.0
pwntools==4.8.0
pycparser==2.21
pyelftools==0.29
Pygments==2.13.0
PyNaCl==1.5.0
pyparsing==3.0.9
pyserial==3.5
PySocks==1.7.1
python-dateutil==2.8.2
pytz==2022.4
pyzmq==24.0.1
rapidfuzz==2.11.1
requests==2.28.1
ROPGadget==7.1
rpyc==5.2.3
seaborn==0.12.0
six==1.16.0
sortedcontainers==2.4.0
soupsieve==2.3.2.post1
SQLAlchemy==1.4.41
stack-data==0.5.1
tornado==6.2
traitlets==5.4.0
unicorn==2.0.0
urllib3==1.26.12
wcwidth==0.2.5
